// src/app/core/models/orders.ts
export class Orders {
    userId?: String;
    firstName?: String;
    lastName?: String;
    email?: String;
    phone?: Number;
    address?: String;
    zipcode?: String;
    city?: String;
    message?: String;
    products?: Array<any>;
    totalQuantity?: Number;
    totalPrice?: Number;
    status?: String;
    date?: Date;
  }